commint of 13-06-2023

+++ src/main/java/org/onLineShop/service/IOrderServiceImpl.java
@@ -7,6 +7,7 @@ import java.util.Date;
 import java.util.List;
 import java.util.Timer;
 import java.util.TimerTask;
+import java.util.stream.Stream;
 
 import org.onLineShop.dao.ItemProductRepository;
 import org.onLineShop.dao.OrdeRepository;
@@ -14,6 +15,7 @@ import org.onLineShop.dao.ProductRepository;
 import org.onLineShop.dao.ShopRepository;
 import org.onLineShop.dao.ShopSellerRepository;
 import org.onLineShop.dao.UserAppRepository;
+import org.onLineShop.entity.Brand;
 import org.onLineShop.entity.EnumStatus;
 import org.onLineShop.entity.ItemProduct;
 import org.onLineShop.entity.Orde;
@@ -25,9 +27,13 @@ import org.onLineShop.service.from.OrderForme;
 import org.onLineShop.service.from.UrlData;
 import org.onLineShop.service.from.UtilDate;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.support.MutableSortDefinition;
+import org.springframework.beans.support.PagedListHolder;
+import org.springframework.beans.support.PropertyComparator;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import jakarta.transaction.Transactional;
@@ -78,6 +84,10 @@ public class IOrderServiceImpl implements IOrderService {
 		orde.setOrderDate(new Date());
 		orde.setStatus(EnumStatus.NOPAID);
 		orde.setOrderPrice(orderForme.getTotalPrince());
+		orde.setName(orderForme.getClient().getName());
+		orde.setAddress(orderForme.getClient().getAddress());
+		orde.setPhoneNumber(orderForme.getClient().getPhoneNumber());
+		orde.setEmail(orderForme.getClient().getEmail());
 		orde.setUserApp(userAppRepository.findById(orderForme.getIdUser()).get());
 		ordeRepository.save(orde);
 		orderForme.getOrdeLine().forEach(ordeLine->{
@@ -110,11 +120,11 @@ public class IOrderServiceImpl implements IOrderService {
 
 	@Override
 	public Page<Orde> getShopOrder(UrlData urlData,PageRequest PageRequest) {
-		if(userAppRepository.findById(urlData.getIdUser()).isPresent()&&shopRepository.findById(urlData.getIdShop()).isEmpty()) {
+		if(userAppRepository.findById(urlData.getIdUser()).isPresent()&&shopRepository.findById(urlData.getIdShop()).isPresent()) {
 			UserApp user = userAppRepository.findById(urlData.getIdUser()).get();
 			Shop shop = shopRepository.findById(urlData.getIdShop()).get();
 			List<Orde> ordes = new ArrayList<>();
-			for(ShopSeller shopSeller:ShopSellerRepository.findAll()) {
+			/*for(ShopSeller shopSeller:ShopSellerRepository.findAll()) {
 				if(shopSeller.getShop().equals(shop)&&shopSeller.getSeller().equals(user)) {
 					for(ItemProduct item: itemProductRepository.findAll()) {
 						if(item.getProduct().getShop().equals(shop)) {
@@ -123,7 +133,32 @@ public class IOrderServiceImpl implements IOrderService {
 					}
 				}
 				return new PageImpl<Orde>(ordes, PageRequest, ordes.size());
-			}
+			}*/
+			/*Stream<ItemProduct> stream = itemProductRepository.findAll().stream();
+			List<ItemProduct> ord =stream.filter(ite->ite.getProduct().getShop().getShopName()=="").toList();*/
+					for(ItemProduct item: itemProductRepository.findAll()) {
+						if(item.getProduct().getShop().equals(shop)) {
+							Boolean find = false;
+							for(Orde orde:ordes) {
+								if(orde.equals(item.getOrde())) {
+									find = true;
+									//break;
+								}
+							}
+							if(!find) {
+								ordes.add(item.getOrde());
+							}
+						}
+					}
+				PagedListHolder<Orde> pagedListHolder = new PagedListHolder<Orde>(ordes);
+				pagedListHolder.setSort(new MutableSortDefinition("orderDate", true,true));
+				pagedListHolder.resort();
+				pagedListHolder.setPage(PageRequest.getPageNumber());
+				pagedListHolder.setPageSize(PageRequest.getPageSize());
+				List<Orde> ordesSlice = pagedListHolder.getPageList();
+				//PropertyComparator.sort(ordesSlice, new MutableSortDefinition("orderDate", true,true));
+				return new PageImpl<Orde>(ordesSlice, PageRequest, ordes.size());
+			
 		}	
 	
 		return null;
@@ -135,12 +170,14 @@ public class IOrderServiceImpl implements IOrderService {
 		TimerTask task = new TimerTask() {
 			@Override
 			public void run() {
-				for(Orde orde:ordeRepository.findAll()) {
-					if(orde.getLimitDate().compareTo(new Date())>0 && orde.getStatus().equals(EnumStatus.NOPAID)) {
-						ordeRepository.delete(orde);
+				if(!ordeRepository.findAll().isEmpty()) {
+					for(Orde orde:ordeRepository.findAll()) {
+						if(orde.getLimitDate().compareTo(new Date())>0 && orde.getStatus().equals(EnumStatus.NOPAID)) {
+							ordeRepository.delete(orde);
+						}
+						
 					}
-					
-				}
+				}	
 							
 			}
